feat: Fix syscall/interrupt handling and improve keyboard driver with semaphore-based blocking
- Fixed INT 0x80 handler to push dummy int_no/error_code before registers to match syscall_regs_t structure layout
- Fixed syscall_entry SYSRET sequence to use temporary register (r10) instead of "pop rsp" which corrupts stack pointer
- Added comprehensive debug logging to Debug Exception handler (INT 0x01) showing RIP/RSP/CS/SS registers
- Migrated keyboard driver from busy-wait polling to semaphore-based blocking in

+++ src/kernel/keymap.h
@@ -4,6 +4,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include <stddef.h>
 
 /* Nombre maximum de keymaps enregistrées */
 #define MAX_KEYMAPS 8
@@ -56,7 +57,7 @@ const keymap_t* keymap_find_by_name(const char* name);
  * @param count Pointeur pour stocker le nombre de keymaps
  * @return Tableau de pointeurs vers les keymaps
  */
-const keymap_t** keymap_list_all(int* count);
+const keymap_t** keymap_list_all(size_t* count);
 
 /**
  * Enregistre une nouvelle keymap.
